jtag_scan: Adjusted the parameter and result types for jtag_scan(), jtag_add_device(), and related functions

+++ src/platforms/hosted/stlinkv2.c
@@ -1029,10 +1029,10 @@ int jtag_scan_stlinkv2(bmp_info_t *info, const uint8_t *irlens)
 		return 0;
 	jtag_dev_count = stlink_read_idcodes(info, idcodes);
 	/* Check for known devices and handle accordingly */
-	for(int i = 0; i < jtag_dev_count; i++)
+	for(uint32_t i = 0; i < jtag_dev_count; i++)
 		jtag_devs[i].jd_idcode = idcodes[i];
-	for(int i = 0; i < jtag_dev_count; i++)
-		for(int j = 0; dev_descr[j].idcode; j++)
+	for(uint32_t i = 0; i < jtag_dev_count; i++)
+		for(size_t j = 0; dev_descr[j].idcode; j++)
 			if((jtag_devs[i].jd_idcode & dev_descr[j].idmask) ==
 			   dev_descr[j].idcode) {
 				if(dev_descr[j].handler)
